authentification files refactoring
