+++ src/__tests__/Login.test.tsx
@@ -1,11 +1,11 @@
 import { render, act, screen } from '@testing-library/react';
 import Home from '../app/login/page';
-import connection from "../app/api/supabase/supabase";
+import connection from "../app/api/supabase/SupabaseContextProvider";
 import userEvent from '@testing-library/user-event';
-import Login from "../app/components/Authentification/Login";
+import Login from "@/app/components/authentification/Login";
 
 // First, mock the necessary parts of Supabase
-jest.mock("../app/api/supabase/supabase", () => ({
+jest.mock("../app/api/supabase/SupabaseContextProvider", () => ({
     auth: {
         signOut: jest.fn(() => Promise.resolve()), // Simulate a successful sign-out
         getSession: jest.fn(() => Promise.resolve({ // Simulate fetching user session
